resourcemanager/commonids: regenerating Management Group ID
