+++ resourcemanager/commonids/management_group.go
@@ -9,16 +9,19 @@ import (
 
 var _ resourceids.ResourceId = ManagementGroupId{}
 
+// ManagementGroupId is a struct representing the Resource ID for a Management Group
 type ManagementGroupId struct {
 	GroupId string
 }
 
+// NewManagementGroupID returns a new ManagementGroupId struct
 func NewManagementGroupID(groupId string) ManagementGroupId {
 	return ManagementGroupId{
 		GroupId: groupId,
 	}
 }
 
+// ParseManagementGroupID parses 'input' into a ManagementGroupId
 func ParseManagementGroupID(input string) (*ManagementGroupId, error) {
 	parser := resourceids.NewParserFromResourceIdType(ManagementGroupId{})
 	parsed, err := parser.Parse(input, false)
@@ -36,6 +39,8 @@ func ParseManagementGroupID(input string) (*ManagementGroupId, error) {
 	return &id, nil
 }
 
+// ParseManagementGroupIDInsensitively parses 'input' case-insensitively into a ManagementGroupId
+// note: this method should only be used for API response data and not user input
 func ParseManagementGroupIDInsensitively(input string) (*ManagementGroupId, error) {
 	parser := resourceids.NewParserFromResourceIdType(ManagementGroupId{})
 	parsed, err := parser.Parse(input, true)
@@ -53,20 +58,38 @@ func ParseManagementGroupIDInsensitively(input string) (*ManagementGroupId, erro
 	return &id, nil
 }
 
+// ValidateManagementGroupID checks that 'input' can be parsed as a Management Group ID
+func ValidateManagementGroupID(input interface{}, key string) (warnings []string, errors []error) {
+	v, ok := input.(string)
+	if !ok {
+		errors = append(errors, fmt.Errorf("expected %q to be a string", key))
+		return
+	}
+
+	if _, err := ParseManagementGroupID(v); err != nil {
+		errors = append(errors, err)
+	}
+
+	return
+}
+
+// ID returns the formatted Management Group ID
 func (id ManagementGroupId) ID() string {
 	fmtString := "/providers/Microsoft.Management/managementGroups/%s"
 	return fmt.Sprintf(fmtString, id.GroupId)
 }
 
+// Segments returns a slice of Resource ID Segments which comprise this Management Group ID
 func (id ManagementGroupId) Segments() []resourceids.Segment {
 	return []resourceids.Segment{
-		resourceids.StaticSegment("providers", "providers", "providers"),
-		resourceids.ResourceProviderSegment("microsoftManagement", "Microsoft.Management", "Microsoft.Management"),
-		resourceids.StaticSegment("managementGroups", "managementGroups", "managementGroups"),
+		resourceids.StaticSegment("staticProviders", "providers", "providers"),
+		resourceids.ResourceProviderSegment("staticMicrosoftManagement", "Microsoft.Management", "Microsoft.Management"),
+		resourceids.StaticSegment("staticManagementGroups", "managementGroups", "managementGroups"),
 		resourceids.UserSpecifiedSegment("groupId", "groupIdValue"),
 	}
 }
 
+// String returns a human-readable description of this Management Group ID
 func (id ManagementGroupId) String() string {
 	components := []string{
 		fmt.Sprintf("Group: %q", id.GroupId),
